Added Images.
This commit contains the basic changes it requires to have Images in the
repository. You'll notice that the Docker::Image class has a LOT in
common with the Docker::Container class. This was intentional, as I find
it easier to abstart the common denominators once all of the information
is present.