Added Docker::Container#run.
This method will allow for some pretty clean Container chaining.
In addition to this method, I also added #start?, #stop?, #kill?,
and #restart methods to the Container which will attempt to run those
methods, but rescue from any Server errors. Essentially, this means it
will try to run those methods unless the Container is in an invalid
state.