Use #streaming_logs instead of #attach in Rspec
In Docker::Image Rspec tests, the actions are happening too fast to successfully attach and collect the STDOUT. Tests were failing because the container in question was already stopped when we tried to attach to it.

In addition, the #streaming_logs method in Docker::Container was appending a newline to the end of each line it was printing. This was causing files to be printed double spaced ("line\n\n"). As such, the array is now being joined on an empty string.