New Feature: Specify repo_tag in #push and #remove
For Docker::Image, this commit adds the ability to specify a specific repo_tag/name to push or remove.

In the API, the push and remove endpoints accept a name, which is different from an ID. The instance methods here were defaulting to just the ID which meant that if two tags pointing to the same image ID we couldn't delete the name (i.e. "untag") or push the repo_tag (i.e. when the same image belongs to two different repositories).