Stub ENV instead of overwriting for docker_spec
The docker_spec tests rely on leveraging the built-in Docker handling. This means that people using boot2docker need to be able to leverage their default configuration. This commit changes the way we handle ENV variables in the docker_spec tests by stubbing the ENV constant instead of overwriting it.

This also introduces a new Docker module method: reset! Like reset_connection!, this endpoint resets the values in Docker.