Update images used in Rspec tests
This commit is kind of large as it entirely changes the base images used in just about every call. The Docker Hub says thats the 'base' image has been deprecated so this is a good opportunity to replace it. The direct replacement would be ubuntu but debian:wheezy has the same necessary feature set for a fraction of the size.

In addition, when images need to be imported and exported, we will use tianon/true (which is the smallest runable Docker image on the Docker Hub). This reduces the need to stub out the HTTP for URI import because we can have a tiny (~ 100B) tarball in the repo. The total size of the spec directory with all the VCRs is ~ 800K.

Anytime we need to download an image over the internet (and not run it) we will use the scratch image. This is the smallest image but it isn't runnable so its perfect for tests when we dont need to have a functioning container.

To facilitate tests where we are creating images that maybe need to be uploaded to a real Docker Hub, we will leverage the DOCKER_API_USER environment variable to help name the images.