Use DOCKER_HOST stub for VCR cassettes
To address the issue where Travis was failing due to the URI not matching. The correct way to fix this issue is to use `:match_requests_on` but the issue is that the `URI` library can not accurately parse unix socket URIs. To get around this we are using the `filter_sensitive_data` feature to replace the host portion of the URI with a stub to be filled at runtime.