Docker::Image#get doesn't have RepoTags
When using Docker::Image#get, the returned results do not have
"RepoTags" - this is apparently by design in the Docker API, but it
makes for some pretty weird behavior in this gem, where the objects
available interface depends on the method that created it.

To work around this in in the Docker::Image#push method, we check to
make sure the `info` hash has they RepoTags key, and if it doesn't we
retrieve it from a different source.