Image#push tests do not cause uncaught unauthorized error
The tests in push took the image tianon/true and did a build with no changes -- giving an image named both `tianon/true` and `<USERNAME>/true`.
Since `tianon/true` came first in the repo tags, the push specs tried to push to tianon/true, who we are not and not, and so the actual recorded response is an unauthorized error.
The next commit will catch such errors, making those tests fail. For cleanliness of diffs, I'm correcting the existing tests first.

My solution is to make the created image non-identical to tianon/true, by adding a gratuitous environment variable to the build script.
That leaves the image at hand named only `<USERNAME>/true`, and so it sails off to your actual account as it should.