Use a different dir to test Image#insert_local
I noticed after adding Chef::Container::Config that doing so broke the Image#insert_local test because there were new files in lib/docker. To avoid this in the future (as well as make the vcr's smaller) I modified the test to upload the spec/fixtures/top directory because its smaller and less likely to change.