`Image.create`: support the `tag` argument
When a `tag` argument is passed to `Image.create`, that argument needs
to be re-used when loading the image after it's pulled, otherwise trying
to get the image will either crash with a not found error, or return the
wrong image (depending on whether the `latest` tag exists locally for
the repository being pulled).

That part is pretty straightforward, but there are some edge cases when
the user is providing a tag with `fromImage` *and* `tag`.

It is probably desirable to throw an error in most cases: if the user is
confused and passing two different tags, `docker-api` probably shouldn't
venture a guess as to which one they *really* meant (although Docker
itself does assume they meant the one from `tag`). Here, this will fail
by trying to look for an image like `busybox:uclibc:glibc`, which Docker
won't find).

However, there is one case where not throwing an error is fine: when
`fromImage` and the `tag` represent the same tag. In this case,
appending the tag would cause an error (we'd be looking for e.g.
`busybox:uclibc:uclibc`), whereas just `get`'ing `image` would do the
right thing. So, I added an exception to not append the tag when that
tag is already present verbatim in the image.