Set the content type header on build
Prior to this patch, the application/json Content-Type header was set
but he body is a tar file as returned by Docker::Util.create_tar. That
this code was there for a long time suggests Docker doesn't check the
content type. However, podman's REST API does check the header and
rejects the request. After this patch, it is accepted.