Automatically retry with correct content-type
The podman socket will return a note about the correct content-type that
should be used for a given type of message if it is incorrect.

This is a workaround to automatically correct the content-type and
resubmit the call.

A full correction would be to update all of the different connection
entries but this appears to work correctly in the short term.

Updated the spec tests to use debian:stable instead of wheezy since
wheezy is no longer widely supported

Added podman-specific skips to spec tests for running on podman systems.

Explicitly use URI.open in image pulls to work around potential issues
with Ruby 2.7+.

Fixes #568